requirements.txt
cryptography==43.0.3
pexpect==4.9.0
python-dotenv==1.0.1
psutil==6.1.0